Introduce not yet working react timeline chart wrapper
