+++ sippy-ng/package.json
@@ -21,8 +21,10 @@
     "chartjs-plugin-trendline": "^0.1.1",
     "chroma-js": "^2.1.2",
     "clsx": "^1.1.1",
+    "d3-scale": "^4.0.2",
     "date-fns": "^2.0.0-beta.5",
     "date-fns-tz": "^1.1.6",
+    "prop-types": "^15.8.1",
     "query-string": "^4.3.4",
     "react": "^17.0.2",
     "react-chartjs-2": "^3.0.4",
