+++ sippy-ng/src/prow_job_runs/ProwJobRun.js
@@ -3,6 +3,7 @@ import { Error } from '@material-ui/icons'
 import Alert from '@material-ui/lab/Alert'
 import PropTypes from 'prop-types'
 import React, { Fragment, useEffect } from 'react'
+import TimelineChart from '../components/TimelineChart'
 
 export default function ProwJobRun(props) {
   const [fetchError, setFetchError] = React.useState('')
@@ -96,12 +97,16 @@ export default function ProwJobRun(props) {
     eventInterval.categories.uncategorized = !_.some(eventInterval.categories) // will save time later during filtering and re-rendering since we don't render any uncategorized events
   })
 
+  let filteredEvents = filterEvents(eventIntervals)
+  let chartData = groupIntervals(filteredEvents)
+
   return (
     /* eslint-disable react/prop-types */
     <Fragment>
-      <p>Loaded {eventIntervals.length} intervals</p>
-
-      <div id="chart"></div>
+      <p>
+        Loaded {eventIntervals.length} intervals. After filtering:{' '}
+        {filteredEvents.length}. Chart data: {chartData.length}
+      </p>
     </Fragment>
   )
 }
@@ -113,7 +118,7 @@ ProwJobRun.propTypes = {
   filterModel: PropTypes.object,
 }
 
-function renderChart() {
+function filterEvents(eventIntervals) {
   let isSet = false
   /*
   let positiveSelectionRows = new Map()
@@ -156,25 +161,18 @@ function renderChart() {
   }
    */
 
-  var filteredEvents
-  // if none of the inputs are set, nothing to filter so don't waste time looping through everything
-  if (!isSet) {
-    filteredEvents = eventIntervals.items
-  } else {
+  // if none of the filter inputs are set, nothing to filter so don't waste time looping through everything
+  let filteredEvents = eventIntervals
+  if (isSet) {
     // At least one of the inputs had a value, test any inputs that had values
     // This currently does an OR operation of the input fields
-    filteredEvents = _.filter(eventIntervals.items, function (eventInterval) {
+    filteredEvents = _.filter(eventIntervals, function (eventInterval) {
       // Go ahead and filter out uncategorized events
       if (eventInterval.categories.uncategorized) {
         return false
       }
 
       let eventMatches = false
-
-      // TODO: remove this temp hack
-      if (eventInterval.categories.alerts) {
-        return true
-      }
 
       /*
       for (let [key, positiveSelectionRow] of positiveSelectionRows) {
@@ -253,6 +251,10 @@ function renderChart() {
     })
   }
 
+  return filteredEvents
+}
+
+function groupIntervals(filteredEvents) {
   let timelineGroups = []
   timelineGroups.push({ group: 'operator-unavailable', data: [] })
   createTimelineData(
@@ -368,135 +370,7 @@ function renderChart() {
     filteredEvents,
     'interesting_events'
   )
-
-  let segmentFunc = function (segment) {
-    // for (var i in data) {
-    //     if (data[i].group == segment.group) {
-    //         var groupdata = data[i].data
-    //         for (var j in groupdata) {
-    //             if (groupdata[j].label == segment.label) {
-    //                 labeldata = groupdata[j].data
-    //                 for (var k in labeldata) {
-    //                     var startDate = new Date(labeldata[k].timeRange[0])
-    //                     var endDate = new Date(labeldata[k].timeRange[1])
-    //                     if (startDate.getTime() == segment.timeRange[0].getTime() &&
-    //                         endDate.getTime() == segment.timeRange[1].getTime()) {
-    //                         $('#myModalContent').text(labeldata[k].extended)
-    //                         $('#myModal').modal()
-    //                     }
-    //                 }
-    //             }
-    //         }
-    //     }
-    // }
-  }
-
-  const el = document.querySelector('#chart')
-  const myChart = TimelinesChart()
-  let ordinalScale = d3
-    .scaleOrdinal()
-    .domain([
-      'InterestingEvent',
-      'PathologicalKnown',
-      'PathologicalNew', // interesting and pathological events
-      'AlertInfo',
-      'AlertPending',
-      'AlertWarning',
-      'AlertCritical', // alerts
-      'OperatorUnavailable',
-      'OperatorDegraded',
-      'OperatorProgressing', // operators
-      'Update',
-      'Drain',
-      'Reboot',
-      'OperatingSystemUpdate',
-      'NodeNotReady', // nodes
-      'Passed',
-      'Skipped',
-      'Flaked',
-      'Failed', // tests
-      'PodCreated',
-      'PodScheduled',
-      'PodTerminating',
-      'ContainerWait',
-      'ContainerStart',
-      'ContainerNotReady',
-      'ContainerReady',
-      'ContainerReadinessFailed',
-      'ContainerReadinessErrored',
-      'StartupProbeFailed', // pods
-      'CIClusterDisruption',
-      'Disruption', // disruption
-      'Degraded',
-      'Upgradeable',
-      'False',
-      'Unknown',
-      'PodLogInfo',
-      'PodLogWarning',
-      'PodLogError',
-    ])
-    .range([
-      '#6E6E6E',
-      '#0000ff',
-      '#d0312d', // pathological and interesting events
-      '#fada5e',
-      '#fada5e',
-      '#ffa500',
-      '#d0312d', // alerts
-      '#d0312d',
-      '#ffa500',
-      '#fada5e', // operators
-      '#1e7bd9',
-      '#4294e6',
-      '#6aaef2',
-      '#96cbff',
-      '#fada5e', // nodes
-      '#3cb043',
-      '#ceba76',
-      '#ffa500',
-      '#d0312d', // tests
-      '#96cbff',
-      '#1e7bd9',
-      '#ffa500',
-      '#ca8dfd',
-      '#9300ff',
-      '#fada5e',
-      '#3cb043',
-      '#d0312d',
-      '#d0312d',
-      '#c90076', // pods
-      '#96cbff',
-      '#d0312d', // disruption
-      '#b65049',
-      '#32b8b6',
-      '#ffffff',
-      '#bbbbbb',
-      '#96cbff',
-      '#fada5e',
-      '#d0312d',
-    ])
-  myChart
-    .data(timelineGroups)
-    .useUtc(true)
-    .zQualitative(true)
-    .enableAnimations(false)
-    .leftMargin(240)
-    .rightMargin(1550)
-    .maxLineHeight(20)
-    .maxHeight(10000)
-    .zColorScale(ordinalScale)
-    .zoomX([
-      new Date(eventIntervals.items[0].from),
-      new Date(eventIntervals.items[eventIntervals.items.length - 1].to),
-    ])
-    .onSegmentClick(segmentFunc)(el)
-
-  // force a minimum width for smaller devices (which otherwise get an unusable display)
-  setTimeout(() => {
-    if (myChart.width() < 3100) {
-      myChart.width(3100)
-    }
-  }, 1)
+  return timelineGroups
 }
 
 function isOperatorAvailable(eventInterval) {
